Move writer/channel/reader to execute.executor package
