+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/controller/SyncTaskController.java
@@ -25,8 +25,8 @@ import info.avalon566.shardingscaling.core.execute.Event;
 import info.avalon566.shardingscaling.core.execute.EventType;
 import info.avalon566.shardingscaling.core.execute.engine.SyncJobExecutor;
 import info.avalon566.shardingscaling.core.execute.engine.local.LocalSyncJobExecutor;
-import info.avalon566.shardingscaling.core.execute.reader.LogPosition;
-import info.avalon566.shardingscaling.core.execute.reader.ReaderFactory;
+import info.avalon566.shardingscaling.core.execute.executor.reader.LogPosition;
+import info.avalon566.shardingscaling.core.execute.executor.reader.ReaderFactory;
 import info.avalon566.shardingscaling.core.execute.util.DataSourceFactory;
 import info.avalon566.shardingscaling.core.execute.util.DbMetaDataUtil;
 import lombok.extern.slf4j.Slf4j;
