+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/execute/metadata/ColumnMetaData.java
@@ -21,6 +21,7 @@ import lombok.Data;
 
 /**
  * Column meta data.
+ *
  * @author avalon566
  */
 @Data
