+++ sharding-scaling-mysql/src/main/java/info/avalon566/shardingscaling/mysql/MySQLJdbcReader.java
@@ -19,7 +19,7 @@ package info.avalon566.shardingscaling.mysql;
 
 import info.avalon566.shardingscaling.core.config.JdbcDataSourceConfiguration;
 import info.avalon566.shardingscaling.core.config.RdbmsConfiguration;
-import info.avalon566.shardingscaling.core.execute.reader.AbstractJdbcReader;
+import info.avalon566.shardingscaling.core.execute.executor.reader.AbstractJdbcReader;
 import info.avalon566.shardingscaling.core.execute.metadata.JdbcUri;
 
 import java.util.List;
