+++ sharding-scaling-postgresql/src/main/java/info/avalon556/shardingscaling/postgresql/PostgreSQLJdbcReader.java
@@ -18,7 +18,7 @@
 package info.avalon556.shardingscaling.postgresql;
 
 import info.avalon566.shardingscaling.core.config.RdbmsConfiguration;
-import info.avalon566.shardingscaling.core.execute.reader.AbstractJdbcReader;
+import info.avalon566.shardingscaling.core.execute.executor.reader.AbstractJdbcReader;
 
 /**
  * PostgreSQL JDBC reader.
