+++ sharding-scaling-postgresql/src/main/java/info/avalon556/shardingscaling/postgresql/PostgreSQLScalingEntry.java
@@ -18,9 +18,9 @@
 package info.avalon556.shardingscaling.postgresql;
 
 import info.avalon566.shardingscaling.core.spi.ScalingEntry;
-import info.avalon566.shardingscaling.core.execute.reader.JdbcReader;
-import info.avalon566.shardingscaling.core.execute.reader.LogReader;
-import info.avalon566.shardingscaling.core.execute.writer.Writer;
+import info.avalon566.shardingscaling.core.execute.executor.reader.JdbcReader;
+import info.avalon566.shardingscaling.core.execute.executor.reader.LogReader;
+import info.avalon566.shardingscaling.core.execute.executor.writer.Writer;
 
 /**
  * PostgreSQL scaling entry.
